Changed the nature of char_dist and its code
char_dist is now an array of Hash objects corresponding to each position. Each
Hash maps a character to the number of occurences it has for that particular
position. The code constructing color_sequences now runs through each
position in each sequence, comparing the character it finds to the Hash
for that position in order to find the appropriate color from color_intervals.

By making char_dist a persistent variable, we can reference it elsewhere
to find the frequency distribution more easily.